requirements.txt
Django==1.3
South==0.7.4
django-tagging==0.3.1
